Move `OutputMode` to `linker.standard._`.
This is part of the plan for the new linker API #2998.
