+++ tools/shared/src/main/scala/org/scalajs/core/tools/linker/standard/OutputMode.scala
@@ -7,7 +7,7 @@
 \*                                                                      */
 
 
-package org.scalajs.core.tools.linker.backend
+package org.scalajs.core.tools.linker.standard
 
 /** JavaScript output mode. */
 sealed abstract class OutputMode
